% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fivethemes-package.R
\docType{package}
\name{fivethemes-package}
\alias{fivethemes-package}
\alias{_PACKAGE}
\title{fivethemes: Minimal ggplot themes}
\description{
Minimal ggplot themes.
}
\author{
\strong{Maintainer}: Ernest Omane-Kodie \email{eokodie@gmail.com}

}
\keyword{internal}
