% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{plot_candlestick}
\alias{plot_candlestick}
\title{plot_candlesticks}
\usage{
plot_candlestick(quotes_tbl, title)
}
\arguments{
\item{quotes_tbl}{A dataframe with OHCL data.}

\item{title}{(string) Plot title.}
}
\description{
plot_candlesticks
}
\keyword{internal}
