% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_5dark}
\alias{theme_5dark}
\title{theme_5dark}
\usage{
theme_5dark(grid = TRUE)
}
\arguments{
\item{grid}{(string) Grid lines. Options include any
combination of "X", "Y", "x" and "y".}
}
\description{
theme_5dark
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, ggplot2::aes(cyl, mpg)) +
  geom_point() +
  annotate(
    "text", x = 4.5, y = 34,
    label = "2019", parse = TRUE
    ) +
  labs(
    title = "Example title",
    subtitle = "Sample subtitle",
    caption = "Example caption"
  )
p + theme_5dark(grid = "Xx")
}
}
