% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_5minimal}
\alias{theme_5minimal}
\title{theme_5minimal}
\usage{
theme_5minimal(grid = "none", bg_fill = "#F3F3F3")
}
\arguments{
\item{grid}{(string) Grid lines. Options include any
combination of "X", "Y", "x" and "y"}

\item{bg_fill}{(string) Plot backgroud file colour.}
}
\description{
theme_5minimal
}
\examples{
\dontrun{
library(ggplot2)
p <- ggplot(mtcars, ggplot2::aes(cyl, mpg)) +
  geom_point() +
  annotate(
    "text", x = 4.5, y = 34,
    label = "2019", parse = TRUE
    ) +
  labs(
    title = "Example title",
    subtitle = "Sample subtitle",
    caption = "Example caption"
  )
p + theme_5minimal()
}

}
