% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_five}
\alias{theme_five}
\title{theme_five}
\usage{
theme_five(grid = "none", ...)
}
\arguments{
\item{grid}{(string) Grid lines. Options include any
combination of "X", "Y", "x" and "y"}

\item{...}{Other arguments passed to ggplot methods.}
}
\description{
theme_five
}
